Align namespaces in extension
