+++ src/Extensions/FipWriter.cs
@@ -9,60 +9,62 @@ using System.ComponentModel;
 using System.Linq;
 using Bonsai.Harp;
 
-[Combinator]
-[DefaultProperty("FileName")]
-[Description("Writes FIP data into a CSV text file.")]
-public class FipWriter : FileSink
+namespace FipExtensions
 {
-    public int? ExpectedRegionCount = null;
-
-    public IObservable<Timestamped<CircleActivityCollection>> Process(IObservable<Timestamped<CircleActivityCollection>> source)
+    [Combinator]
+    [DefaultProperty("FileName")]
+    [Description("Writes FIP data into a CSV text file.")]
+    public class FipWriter : FileSink
     {
-        var sink = new FipCsvWriter(this, ExpectedRegionCount)
+        public int? ExpectedRegionCount = null;
+
+        public IObservable<Timestamped<CircleActivityCollection>> Process(IObservable<Timestamped<CircleActivityCollection>> source)
         {
-            FileName = FileName,
-            Suffix = Suffix,
-            Buffered = Buffered,
-            Overwrite = Overwrite,
-        };
-        return sink.Process(source);
-    }
-
-    class FipCsvWriter : FileSink<Timestamped<CircleActivityCollection>, StreamWriter>
-    {
-
-        internal int? ExpectedRegionCount = null;
-        const int MetadataOffset = 3;
-
-        internal FipCsvWriter(FipWriter writer, int? expectedRegionCount)
-        {
-            Writer = writer;
-            ExpectedRegionCount = expectedRegionCount;
+            var sink = new FipCsvWriter(this, ExpectedRegionCount)
+            {
+                FileName = FileName,
+                Suffix = Suffix,
+                Buffered = Buffered,
+                Overwrite = Overwrite,
+            };
+            return sink.Process(source);
         }
 
-        internal FipWriter Writer { get; private set; }
-
-        protected override StreamWriter CreateWriter(string fileName, Timestamped<CircleActivityCollection> input)
+        class FipCsvWriter : FileSink<Timestamped<CircleActivityCollection>, StreamWriter>
         {
-            if (Path.GetExtension(fileName) != ".csv")
+
+            internal int? ExpectedRegionCount = null;
+            const int MetadataOffset = 3;
+
+            internal FipCsvWriter(FipWriter writer, int? expectedRegionCount)
             {
-                throw new ArgumentException("File extension must be .csv");
+                Writer = writer;
+                ExpectedRegionCount = expectedRegionCount;
             }
 
-            var nRegions = input.Value.Count;
+            internal FipWriter Writer { get; private set; }
 
-            if (ExpectedRegionCount == null)
+            protected override StreamWriter CreateWriter(string fileName, Timestamped<CircleActivityCollection> input)
             {
-                ExpectedRegionCount = nRegions;
-            }
-            else if (ExpectedRegionCount != nRegions)
-            {
-                throw new ArgumentException("Number of regions in the input stream does not match the number of regions in the first frame.");
-            }
+                if (Path.GetExtension(fileName) != ".csv")
+                {
+                    throw new ArgumentException("File extension must be .csv");
+                }
 
-            if (nRegions == 0) throw new ArgumentException("No regions defined for FipWriter.");
-            var writer = new StreamWriter(fileName, false, Encoding.ASCII);
-            var columns = new List<string>(MetadataOffset + nRegions)
+                var nRegions = input.Value.Count;
+
+                if (ExpectedRegionCount == null)
+                {
+                    ExpectedRegionCount = nRegions;
+                }
+                else if (ExpectedRegionCount != nRegions)
+                {
+                    throw new ArgumentException("Number of regions in the input stream does not match the number of regions in the first frame.");
+                }
+
+                if (nRegions == 0) throw new ArgumentException("No regions defined for FipWriter.");
+                var writer = new StreamWriter(fileName, false, Encoding.ASCII);
+                var columns = new List<string>(MetadataOffset + nRegions)
             {
                 //columns.Add(nameof(input.FrameCounter));
                 //columns.Add(nameof(input.Timestamp));
@@ -72,40 +74,41 @@ public class FipWriter : FileSink
                 "Background" // We assume that the first region is always the background
             };
 
-            if (nRegions > 0)
-            {
-                for (int i = 1; i < nRegions; i++)
+                if (nRegions > 0)
                 {
-                    columns.Add("Region" + i.ToString(CultureInfo.InvariantCulture));
+                    for (int i = 1; i < nRegions; i++)
+                    {
+                        columns.Add("Region" + i.ToString(CultureInfo.InvariantCulture));
+                    }
                 }
+                var header = string.Join(",", columns);
+                writer.WriteLine(header);
+                return writer;
             }
-            var header = string.Join(",", columns);
-            writer.WriteLine(header);
-            return writer;
-        }
 
-        protected override void Write(StreamWriter writer, Timestamped<CircleActivityCollection> input)
-        {
-            var nRegions = input.Value.Count;
-
-            if (nRegions != ExpectedRegionCount)
+            protected override void Write(StreamWriter writer, Timestamped<CircleActivityCollection> input)
             {
-                throw new ArgumentException("Number of regions in the input stream does not match the number of regions in the first frame.");
-            }
-            var values = new List<string>(MetadataOffset + nRegions);
+                var nRegions = input.Value.Count;
 
-            for (int i = 0; i < MetadataOffset; i++)
-            {
-                values.Add("DummyMetadata_" + i.ToString(CultureInfo.InvariantCulture));
-            }
+                if (nRegions != ExpectedRegionCount)
+                {
+                    throw new ArgumentException("Number of regions in the input stream does not match the number of regions in the first frame.");
+                }
+                var values = new List<string>(MetadataOffset + nRegions);
 
-            var activity = input.Value.Select(x => x.Activity).ToArray();
+                for (int i = 0; i < MetadataOffset; i++)
+                {
+                    values.Add("DummyMetadata_" + i.ToString(CultureInfo.InvariantCulture));
+                }
 
-            for (int i = 0; i < activity.Length; i++)
-            {
-                values.Add(activity[i].Val0.ToString(CultureInfo.InvariantCulture));
+                var activity = input.Value.Select(x => x.Activity).ToArray();
+
+                for (int i = 0; i < activity.Length; i++)
+                {
+                    values.Add(activity[i].Val0.ToString(CultureInfo.InvariantCulture));
+                }
+                var line = string.Join(",", values);
             }
-            var line = string.Join(",", values);
         }
     }
 }
